+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 2021.2
 
 * I switched to the .NET 5.0 framework
+* I added some properties to the LinqRepositoryOptions type
 
 ## 2021.1